requirements.txt
huggingface-hub==0.14.1
importlib-metadata==6.6.0
numpy
opencv-python==4.4.0.42
packaging==23.1
pandas==1.1.5
Pillow==9.5.0
six==1.16.0
timm==0.6.13
tokenizers==0.13.3
tqdm==4.65.0
transformers==4.28.1
fairscale==0.4.6
wandb
